Update Actual And Estimated costs for orchestrator
Changed the EstimationOrchestrator to show the Actual and Estimated
costs. Actual costs are a sum of the ActualCommitedTransfer Costs and
CompletionCosts, the EstimatedCost is the sum of the TransferOutCosts
and Completion Costs.

Added toList for the TransferAllowances to stop display errors when
iterating through the collection for the CostEstimation view.